SPlit application controller into separate files for separate concerns
