+++ Superhero/src/main/java/superhero/controller/HomePageController.java
@@ -1,0 +1,19 @@
+package superhero.controller;
+
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api")
+public class HomePageController {
+
+    @GetMapping
+    public String getHomePage() {
+        return "NOT IMPLEMENTED: Send homepage";
+    }
+}
